refactor(frontend): extract area-validation dialog state into useAreaValidationDialog
Move the planting-plan area-validation dialog's state, suppression flag
and close-suppression timer lifecycle out of PlantingPlans.tsx into a
useAreaValidationDialog hook (with the AreaValidationDialogState type and
the suppression-window constant). The page consumes the returned state,
refs and open/close callbacks; the save-guard and commit handler read the
exposed refs exactly as before. No behavior changes.
